Refactor powerups
Players now have a Powerup powerup field, which represents current
powerup. I think you'll agree that this works way better than an int,
as there are no more switch statemtents in Player or GameSubpanel.

Powerups are drawn on screen and in the status panel according to their
color field.

Powerups have a getAmmoType which returns the class that corresponds
to the ammo they fire. If this is null, the Powerup will not change
the ammo type.

Players check the ammo type of their current powerup to determine what
kind of bullet to fire (then create the bullet with some fancy
reflection).